rag_hack/api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
gunicorn==23.0.0
pydantic==2.8.2
coloredlogs==15.0.1
pandas==2.2.2
openai==1.43.0
azure-cognitiveservices-speech==1.32.1
pydub==0.25.1
moviepy==1.0.3
azure-identity==1.17.1
azure-storage-blob==12.22.0
langchain
langchain-openai
langchain-core
langchain-community
python-dotenv
transformers==4.44.2
azure-search-documents==11.4.0
Pillow==9.4.0
torch==2.4.0
python-multipart==0.0.9
neo4j==5.24.0